improve code

+++ app/Http/Controllers/Api/BillAttachmentController.php
@@ -8,11 +8,17 @@ use App\Http\Requests\Api\UpdateAttachmentRequest;
 use App\Http\Resources\AttachmentResource;
 use App\Models\Attachment;
 use App\Models\Bill;
+use Illuminate\Support\Facades\Gate;
 
 class BillAttachmentController extends Controller
 {
     public function index(Bill $bill)
     {
+
+        $user = auth()->user();
+
+        Gate::authorize('viewAny', [$bill, $user]);
+
         $attachments = $bill->attachments()->paginate(10);
 
         return AttachmentResource::collection($attachments);
@@ -20,6 +26,10 @@ class BillAttachmentController extends Controller
 
     public function store(StoreAttachmentRequest $request, Bill $bill)
     {
+        $user = auth()->user();
+
+        Gate::authorize('create', [$bill, $user]);
+
         $file_path = $request->file('file')->store('attachments');
 
         $request->merge(['file_path' => $file_path]);
@@ -31,11 +41,21 @@ class BillAttachmentController extends Controller
 
     public function show(Attachment $attachment)
     {
+
+        $user = auth()->user();
+
+        Gate::authorize('view', [$attachment, $user]);
+
         return new AttachmentResource($attachment);
     }
 
     public function update(UpdateAttachmentRequest $request, Attachment $attachment)
     {
+
+        $user = auth()->user();
+
+        Gate::authorize('update', [$attachment, $user]);
+
         $attachment->update($request->all());
 
         return new AttachmentResource($attachment);
@@ -43,6 +63,11 @@ class BillAttachmentController extends Controller
 
     public function destroy(Attachment $attachment)
     {
+
+        $user = auth()->user();
+
+        Gate::authorize('delete', [$attachment, $user]);
+        
         $attachment->delete();
 
         return response()->noContent();
